Update in ticket label delete and edit also fix xss in add label: #1

+++ Controller/TicketXHR.php
@@ -86,12 +86,14 @@ class TicketXHR extends AbstractController
         $content = $request->getContent();
         $em = $this->getDoctrine()->getManager();
 
-        if($method == "POST") {
+        if ($method == "POST") {
             $data = json_decode($content, true);
-            if($data['name'] != "") {
+
+            if ($data['name'] != "") {
                 $label = new SupportLabel();
-                $label->setName($data['name']);
-                if(isset($data['colorCode']))
+                $label->setName(htmlspecialchars($data['name']));
+                
+                if (isset($data['colorCode']))
                     $label->setColorCode($data['colorCode']);
                 $label->setUser($this->userService->getCurrentUser());
                 $em->persist($label);
@@ -109,12 +111,18 @@ class TicketXHR extends AbstractController
                 $json['alertClass'] = 'danger';
                 $json['alertMessage'] = $this->translator->trans('Error ! Label name can not be blank.');
             }
-        } elseif($method == "PUT") {
+        } elseif ($method == "PUT") {
             $data = json_decode($content, true);
             $label = $em->getRepository(SupportLabel::class)->findOneBy(array('id' => $request->attributes->get('ticketLabelId')));
-            if($label) {
-                $label->setName($data['name']);
-                if(!empty($data['colorCode'])) {
+            
+            if ($label->getUser()->getId() != $this->getUser()->getId()) {
+                throw new \Exception('Access Denied', 403);
+            }
+
+            if ($label) {
+                $label->setName(htmlspecialchars($data['name']));
+               
+                if (!empty($data['colorCode'])) {
                     $label->setColorCode($data['colorCode']);
                 }
                 $em->persist($label);
@@ -132,9 +140,14 @@ class TicketXHR extends AbstractController
                 $json['alertClass'] = 'danger';
                 $json['alertMessage'] = $this->translator->trans('Error ! Invalid label id.');
             }
-        } elseif($method == "DELETE") {
+        } elseif ($method == "DELETE") {
             $label = $em->getRepository(SupportLabel::class)->findOneBy(array('id' => $request->attributes->get('ticketLabelId')));
-            if($label) {
+            
+            if ($label->getUser()->getId() != $this->getUser()->getId()) {
+                throw new \Exception('Access Denied', 403);
+            }
+
+            if ($label) {
                 $em->remove($label);
                 $em->flush();
                 $json['alertClass'] = 'success';
@@ -220,6 +233,7 @@ class TicketXHR extends AbstractController
                 'alertClass' => 'danger',
                 'alertMessage' => $this->translator->trans('Insufficient details provided.'),
             ];
+
             return new Response(json_encode($responseContent), 400, ['Content-Type' => 'application/json']);
         }
 
@@ -561,6 +575,7 @@ class TicketXHR extends AbstractController
         if ('POST' == $request->getMethod()) {
             $responseContent = [];
             $user = $this->userService->getSessionUser();
+            
             $supportLabel = $entityManager->getRepository(SupportLabel::class)->findOneBy([
                 'user' => $user->getId(),
                 'name' => $requestContent['name'],
@@ -568,7 +583,7 @@ class TicketXHR extends AbstractController
 
             if (empty($supportLabel)) {
                 $supportLabel = new SupportLabel();
-                $supportLabel->setName($requestContent['name']);
+                $supportLabel->setName(htmlspecialchars($requestContent['name']));
                 $supportLabel->setUser($user);
 
                 $entityManager->persist($supportLabel);
@@ -589,6 +604,7 @@ class TicketXHR extends AbstractController
                 ]);
             } else {
                 $isLabelAlreadyAdded = false;
+                
                 foreach ($ticketLabelCollection as $ticketLabel) {
                     if ($supportLabel->getId() == $ticketLabel->getId()) {
                         $isLabelAlreadyAdded = true;
@@ -640,7 +656,7 @@ class TicketXHR extends AbstractController
             ->setParameter('labelUserId', $this->getUser()->getId())
             ->setParameter('companyId', $this->getCompany()->getId());
 
-        if($request) {
+        if ($request) {
             $qb->andwhere("tl.name LIKE :labelName");
             $qb->setParameter('labelName', '%'.urldecode($request->query->get('query')).'%');
         }
@@ -818,11 +834,11 @@ class TicketXHR extends AbstractController
             throw new \Exception('Access Denied', 403);
         }
 
-        if($request->getMethod() == "POST") {
+        if ($request->getMethod() == "POST") {
             $tag = new CoreFrameworkBundleEntities\Tag();
             if ($content['name'] != "") {
                 $checkTag = $em->getRepository(Tag::class)->findOneBy(array('name' => $content['name']));
-                if(!$checkTag) {
+                if (!$checkTag) {
                     $tag->setName($content['name']);
                     $em->persist($tag);
                     $em->flush();
@@ -840,11 +856,11 @@ class TicketXHR extends AbstractController
                 $json['alertClass'] = 'danger';
                 $json['alertMessage'] = $this->translator->trans('Please enter tag name.');
             }
-        } elseif($request->getMethod() == "DELETE") {
+        } elseif ($request->getMethod() == "DELETE") {
             $tag = $em->getRepository(Tag::class)->findOneBy(array('id' => $request->attributes->get('id')));
-            if($tag) {
+            if ($tag) {
                 $articles = $em->getRepository(ArticleTags::class)->findOneBy(array('tagId' => $tag->getId()));
-                if($articles)
+                if ($articles)
                     foreach ($articles as $entry) {
                         $em->remove($entry);
                     }
@@ -870,17 +886,17 @@ class TicketXHR extends AbstractController
     {
         $json = [];
         if ($request->isXmlHttpRequest()) {
-            if($request->query->get('type') == 'agent') {
+            if ($request->query->get('type') == 'agent') {
                 $json = $this->userService->getAgentsPartialDetails($request);
-            } elseif($request->query->get('type') == 'customer') {
+            } elseif ($request->query->get('type') == 'customer') {
                 $json = $this->userService->getCustomersPartial($request);
-            } elseif($request->query->get('type') == 'group') {
+            } elseif ($request->query->get('type') == 'group') {
                 $json = $this->userService->getSupportGroups($request);
-            } elseif($request->query->get('type') == 'team') {
+            } elseif ($request->query->get('type') == 'team') {
                 $json = $this->userService->getSupportTeams($request);
-            } elseif($request->query->get('type') == 'tag') {
+            } elseif ($request->query->get('type') == 'tag') {
                 $json = $this->ticketService->getTicketTags($request);
-            } elseif($request->query->get('type') == 'label') {
+            } elseif ($request->query->get('type') == 'label') {
                 $json = $this->ticketService->getLabels($request);
             }
         }
@@ -903,7 +919,7 @@ class TicketXHR extends AbstractController
         }
 
         if($request->getMethod() == "POST") {
-            if($content['email'] == $ticket->getCustomer()->getEmail()) {
+            if ($content['email'] == $ticket->getCustomer()->getEmail()) {
                 $json['alertClass'] = 'danger';
                 $json['alertMessage'] = $this->translator->trans('Error ! Customer can not be added as collaborator.');
             } else {
@@ -947,9 +963,9 @@ class TicketXHR extends AbstractController
                     $json['alertMessage'] = $this->translator->trans('Error ! ' . $message);
                 }
             }
-        } elseif($request->getMethod() == "DELETE") {
+        } elseif ($request->getMethod() == "DELETE") {
             $collaborator = $em->getRepository(User::class)->findOneBy(array('id' => $request->attributes->get('id')));
-            if($collaborator) {
+            if ($collaborator) {
                 $ticket->removeCollaborator($collaborator);
                 $em->persist($ticket);
                 $em->flush();
@@ -989,7 +1005,7 @@ class TicketXHR extends AbstractController
 
         if (isset($content['name']) && $content['name'] != "") {
             $checkTag = $entityManager->getRepository(Tag::class)->findOneBy(array('id' => $tagId));
-            if($checkTag) {
+            if ($checkTag) {
                 $checkTag->setName($content['name']);
                 $entityManager->persist($checkTag);
                 $entityManager->flush();
@@ -1009,7 +1025,7 @@ class TicketXHR extends AbstractController
         $entityManager = $this->getDoctrine()->getManager();
         $checkTag = $entityManager->getRepository(Tag::class)->findOneBy(array('id' => $tagId));
 
-        if($checkTag) {
+        if ($checkTag) {
             $entityManager->remove($checkTag);
             $entityManager->flush();
